Use Standard AWS Credential Process
Since there is *always* an `awsprofile` name passed, either the
one from the command line or `'default'`, it is impossible to use
environment variables like `AWS_PROFILE`, `AWS_ACCESS_KEY_ID`,
`AWS_SECRET_ACCESS_KEY` etc. because it instructs boto3 to always look
for a named profile.

By making the profile name optional, we keep the same external behavior
(since the default profile name is `default` as well and chosen if none
of the above is given), but enable us to use the standard credential
process of boto(core).

https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-profiles.html